requirements.txt
aiofiles==23.2.1
altair==5.3.0
anyio==4.3.0
asttokens==2.4.1
attrs==23.2.0
beautifulsoup4==4.12.3
blinker==1.8.1
branca==0.7.2
Brotli @ file:///C:/Windows/Temp/abs_63l7912z0e/croots/recipe/brotli-split_1659616056886/work
cachetools==5.3.3
certifi @ file:///C:/b/abs_91u83siphd/croot/certifi_1700501720658/work/certifi
cffi @ file:///C:/b/abs_924gv1kxzj/croot/cffi_1700254355075/work
chardet==5.2.0
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
ci-info==0.3.0
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
cloudpickle==3.0.0
colorama==0.4.6
comm==0.2.1
configobj==5.0.8
configparser==7.0.0
contourpy==1.2.0
cryptography @ file:///C:/b/abs_e8cnom_zw_/croot/cryptography_1702071486468/work
cycler==0.12.1
Cython==0.29.37
debugpy==1.8.0
decorator==5.1.1
etelemetry==0.3.1
exceptiongroup==1.2.0
executing==2.0.1
filelock @ file:///C:/b/abs_f2gie28u58/croot/filelock_1700591233643/work
fiona==1.9.6
Flask==3.0.3
Flask-Cors==4.0.1
folium==0.16.0
fonttools==4.47.2
frontend==0.0.3
fsspec==2024.3.1
geopandas==0.14.3
gitdb==4.0.11
GitPython==3.1.43
gmpy2 @ file:///C:/ci/gmpy2_1645438895476/work
gym==0.26.2
gym-notices==0.0.8
h11==0.14.0
hdbscan==0.8.33
httplib2==0.22.0
huggingface-hub==0.22.2
idna @ file:///C:/b/abs_bdhbebrioa/croot/idna_1666125572046/work
imageio==2.33.1
ipykernel==6.29.0
ipython==8.21.0
isodate==0.6.1
itsdangerous==2.2.0
jedi==0.19.1
Jinja2 @ file:///C:/b/abs_7cdis66kl9/croot/jinja2_1666908141852/work
joblib==1.3.2
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.0
jupyter_core==5.7.1
kiwisolver==1.4.5
lazy_loader==0.3
llvmlite==0.42.0
looseversion==1.3.0
lxml==5.2.1
markdown-it-py==3.0.0
MarkupSafe @ file:///C:/b/abs_ecfdqh67b_/croot/markupsafe_1704206030535/work
matplotlib==3.8.2
matplotlib-inline==0.1.6
mdurl==0.1.2
mkl-fft @ file:///C:/b/abs_19i1y8ykas/croot/mkl_fft_1695058226480/work
mkl-random @ file:///C:/b/abs_edwkj1_o69/croot/mkl_random_1695059866750/work
mkl-service==2.4.0
mpmath @ file:///C:/b/abs_7833jrbiox/croot/mpmath_1690848321154/work
nest-asyncio==1.6.0
networkx @ file:///C:/b/abs_e6gi1go5op/croot/networkx_1690562046966/work
nibabel==5.2.1
nipype==1.8.6
numba==0.59.1
numpy @ file:///C:/b/abs_16b2j7ad8n/croot/numpy_and_numpy_base_1704311752418/work/dist/numpy-1.26.3-cp310-cp310-win_amd64.whl#sha256=e84057072c37569bd0e11652dc2a75980d4d360f2391adf6a29a2fb1622d20ff
packaging==23.2
pandas==2.2.2
parso==0.8.3
pathlib==1.0.1
patsy==0.5.6
Pillow @ file:///C:/b/abs_20ztcm8lgk/croot/pillow_1696580089746/work
platformdirs==4.2.0
prompt-toolkit==3.0.43
protobuf==4.25.3
prov==2.0.0
psutil==5.9.8
pure-eval==0.2.2
pyarrow==16.0.0
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pydeck==0.9.1
pydot==2.0.0
Pygments==2.17.2
PyMuPDF==1.24.3
PyMuPDFb==1.24.3
pynndescent==0.5.12
pyOpenSSL @ file:///C:/b/abs_08f38zyck4/croot/pyopenssl_1690225407403/work
pyparsing==3.1.1
pyproj==3.6.1
PySocks @ file:///C:/ci_310/pysocks_1642089375450/work
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2024.1
pywin32==306
pyxnat==1.6.2
PyYAML @ file:///C:/b/abs_782o3mbw7z/croot/pyyaml_1698096085010/work
pyzmq==25.1.2
rdflib==7.0.0
referencing==0.35.1
regex==2024.4.16
requests @ file:///C:/b/abs_316c2inijk/croot/requests_1690400295842/work
rich==13.7.1
rpds-py==0.18.1
safetensors==0.4.3
scikit-image==0.22.0
scikit-learn==1.4.0
scipy==1.12.0
seaborn==0.13.2
sentence-transformers==2.7.0
shapely==2.0.4
simplejson==3.19.2
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.5
stack-data==0.6.3
starlette==0.37.2
statsmodels==0.14.2
streamlit==1.34.0
sympy @ file:///C:/b/abs_82njkonm7f/croot/sympy_1701397685028/work
tenacity==8.3.0
threadpoolctl==3.2.0
tifffile==2024.1.30
tokenizers==0.19.1
toml==0.10.2
toolz==0.12.1
torch==2.2.0
torchaudio==2.2.0
torchvision==0.17.0
tornado==6.4
tqdm==4.66.1
traitlets==5.14.1
traits==6.3.2
transformers==4.40.1
typing_extensions @ file:///C:/b/abs_72cdotwc_6/croot/typing_extensions_1705599364138/work
tzdata==2024.1
umap-learn==0.5.6
urllib3 @ file:///C:/b/abs_9cmlsrm3ys/croot/urllib3_1698257595508/work
uvicorn==0.29.0
watchdog==4.0.0
wcwidth==0.2.13
Werkzeug==3.0.2
wikipedia==1.4.0
win-inet-pton @ file:///C:/ci_310/win_inet_pton_1642658466512/work
xyzservices==2024.4.0
